Tool V2/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=1.0.0
plotly>=5.14.0
scikit-learn>=1.2.0
pytest>=7.0.0
